feat(timepicker): fix selection state for option
